make product locator optional on detail page
on detail page we have an optional back and next link, but the the current position of the article was not optional in the template.
For us it is important to have this optional because it is not always possible to locate the article (sometimes not simple without loosing performance). Without this change you sometimes get a broken view.