PR-132 Fix errors in translations
https://bugs.oxid-esales.com/view.php?id=6909
Proposals by pbenke.

I've struggled to decide whether "E-mail" or "Email" is correct. But I've got from four different translation sites the information that "Email" is more or less a synonym of "E-mail" as "electronic mail" are two words and the E is just a abbreviation for electronic. So it must chained with a minus.